Move DebugSession to its own module.
